Move Semantic Kernel plugins into its own project
