+++ AIDemo.sln
@@ -51,11 +51,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.ServiceDefaults", "s
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.AppHost", "src\AIDemo.AppHost\AIDemo.AppHost.csproj", "{6A5CC921-796B-4D16-84DD-0E105B48BE27}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.Messages", "src\AIDemo.Messages\AIDemo.Messages.csproj", "{2B8963B6-E1C0-48FF-8A11-9CC4A23785B5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.Contracts", "src\AIDemo.Contracts\AIDemo.Contracts.csproj", "{2B8963B6-E1C0-48FF-8A11-9CC4A23785B5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.Entities", "src\AIDemo.Entities\AIDemo.Entities.csproj", "{F7EEFDD5-A299-4F79-9BAD-6199FC385910}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.Hubs", "src\AIDemo.Hubs\AIDemo.Hubs.csproj", "{7C955F80-4473-4F59-986F-99E3EA82F677}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIDemo.SemanticKernel", "src\AIDemo.SemanticKernel\AIDemo.SemanticKernel.csproj", "{CEDC78FB-5DB8-4F64-9196-9517F4C6DF23}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -94,6 +96,10 @@ Global
 		{7C955F80-4473-4F59-986F-99E3EA82F677}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C955F80-4473-4F59-986F-99E3EA82F677}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7C955F80-4473-4F59-986F-99E3EA82F677}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CEDC78FB-5DB8-4F64-9196-9517F4C6DF23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CEDC78FB-5DB8-4F64-9196-9517F4C6DF23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CEDC78FB-5DB8-4F64-9196-9517F4C6DF23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CEDC78FB-5DB8-4F64-9196-9517F4C6DF23}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{367AFC48-B548-40A9-9DEC-B56979EC7082} = {7FCDD3F2-5976-443D-B5C2-AC71C062C0FB}
