+++ src/AIDemo.Contracts/Clients/IGoogleGeocodeClient.cs
