+++ src/AIDemo.Contracts/Clients/IWeatherApiClient.cs
