+++ src/AIDemo.Contracts/Config/GitHubOptions.cs
