+++ src/AIDemo.Contracts/Config/GoogleOptions.cs
