+++ src/AIDemo.Contracts/Config/WeatherOptions.cs
