+++ src/AIDemo.Contracts/DistanceCalculator.cs
