+++ src/AIDemo.Contracts/Expect.cs
