+++ src/AIDemo.Contracts/Extensions/StringExtensions.cs
