+++ src/AIDemo.Contracts/Geocoding.cs
