+++ src/AIDemo.Contracts/Messages.cs
