+++ src/AIDemo.Contracts/Models.cs
