+++ src/AIDemo.Contracts/Results.cs
