+++ src/AIDemo.Entities/AIDemo.Entities.csproj
@@ -13,7 +13,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\AIDemo.Messages\AIDemo.Messages.csproj" />
+      <ProjectReference Include="..\AIDemo.Contracts\AIDemo.Contracts.csproj" />
     </ItemGroup>
 
 </Project>
