+++ src/AIDemo.SemanticKernel/FunctionSignalFilter.cs
