+++ src/AIDemo.SemanticKernel/HubContextExtensions.cs
