+++ src/AIDemo.SemanticKernel/Plugins/ContactFactsPlugin.cs
