+++ src/AIDemo.SemanticKernel/Plugins/ContactPlugin.cs
