+++ src/AIDemo.SemanticKernel/Plugins/LocationFunction.cs
