+++ src/AIDemo.SemanticKernel/Plugins/WeatherPlugin.cs
