+++ src/AIDemo.Web/AIDemo.Web.csproj
@@ -68,7 +68,8 @@
   <ItemGroup>
     <ProjectReference Include="..\AIDemo.Entities\AIDemo.Entities.csproj" />
     <ProjectReference Include="..\AIDemo.Hubs\AIDemo.Hubs.csproj" />
-    <ProjectReference Include="..\AIDemo.Messages\AIDemo.Messages.csproj" />
+    <ProjectReference Include="..\AIDemo.Contracts\AIDemo.Contracts.csproj" />
+    <ProjectReference Include="..\AIDemo.SemanticKernel\AIDemo.SemanticKernel.csproj" />
     <ProjectReference Include="..\AIDemo.ServiceDefaults\AIDemo.ServiceDefaults.csproj" />
   </ItemGroup>
 </Project>
