+++ src/AIDemo.Web/Program.cs
@@ -4,6 +4,7 @@ using Haack.AIDemoWeb.Components; // Rider highlights this line for some reason,
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using AIDemo.Hubs;
+using Haack.AIDemoWeb.Library;
 
 var builder = WebApplication.CreateBuilder(args);
